database and server connected. protection layer added.
