+++ db/index.js
@@ -1,49 +1,25 @@
-const Sequelize = require('sequelize');
+const { Pool } = require('pg');
 
-const sequelize = new Sequelize('booking', 'root', null, {
-  host: 'localhost',
-  dialect: 'mysql',
+const pool = new Pool({
+  database: 'listings',
 });
 
-
-const Room = sequelize.define('rooms', {
-  id: { type: Sequelize.INTEGER, primaryKey: true, autoIncrement: true },
-  roomname: Sequelize.STRING,
-  price: Sequelize.INTEGER,
-  cleaning_fee: Sequelize.INTEGER,
-  service_fee: Sequelize.INTEGER,
-  tax: Sequelize.INTEGER,
-  max_guest: Sequelize.STRING,
-  min_night: Sequelize.INTEGER,
-  max_night: Sequelize.INTEGER,
-  ratings: Sequelize.DECIMAL(2, 1),
-  num_reviews: Sequelize.INTEGER,
+pool.connect(() => {
+  console.log('connected to PostgreSQL');
 });
 
-const Booking = sequelize.define('bookings', {
-  email: Sequelize.STRING,
-  guests: Sequelize.STRING,
-  check_in: Sequelize.DATE,
-  check_out: Sequelize.DATE,
-  createdAt: Sequelize.DATE,
-  roomId: {
-    type: Sequelize.INTEGER,
-    references: {
-      model: 'rooms',
-      key: 'id',
-    },
-  },
-});
-
-Room.hasMany(Booking, { foreignKey: 'roomId' });
-Booking.belongsTo(Room, { foreignKey: 'roomId' });
-
-sequelize.authenticate();
-
-Room.sync();
-Booking.sync();
+const getRoom = (id, callback) => {
+  const query = `SELECT * FROM rooms WHERE id = ${Number(id)}`;
+  pool.query(query)
+    .then((res) => {
+      callback(null, res);
+    })
+    .catch((err) => {
+      callback(err);
+    });
+};
 
 module.exports = {
-  Room,
-  Booking,
+  pool,
+  getRoom,
 };
